add hosts setting
app.config['DEBUG_TB_HOSTS']

Example: app.config['DEBUG_TB_HOSTS'] = ( '127.0.0.1', '::1' )

If the request’s REMOTE_ADDR is not in this list, the toolbar will not
be displayed and the exception handler will not be active. This should
be a tuple. Default: (). Empty list remains backward compatibility. If
DEBUG_TB_HOSTS not specified in app.config, the debug toolbar is shown
for all addresses.