Bump `jquery` to `3.7.1` and `jquery-migrate` to `3.4.1`
Bump to the latest versions of `jquery` and `jquery-migrate`.

I popped open the browser console and I see five deprecation warnings,
but 3 out of the 5 are triggered by `jquery.tablesorter.js` which we
vendor, so nothing we can do there.

The other two we should address, but we can do so in a follow-on PR as
they are merely deprecations at this point, not fatal errors.

So getting onto the `3.x` series, even with deprecation warnings, is a
nice step forward.

Previously we shipped the minified version of `jquery`, and we could
continue to do so, but given this is a dev tool and we still have
warnings we need to fix, I opted for the non-minified version for now.

Additionally, they offer a `slim` varient that is stripped down... we
could probably use that too, but I didn't take the time to investigate.