Install custom Sphinx theme via `extras_require`
Previously the custom theme was vendored in via a `git` submodule...

Now this theme is available via a python package, so install it using
`extras_require` instead.

This also unlocks letting other tooling such as ReadTheDocs install the
custom theme without having to do custom git incantations.

I also removed the `index_logo` config as Sphinx warns that it's not a
supported option by the theme. I grep'd the theme package, and it makes
no mention of this option, besides we were leaving it blank already, so
there's no point in having it.